Create Function-procedure-with-return-value.java
 It is better to write functions which return a value and do not display inside
the function, because the function is more likely to be useful at other times without
changes. If the procedure is to be used elsewhere it is better if it does not access a global
variable, because the variable name the procedure uses may already be used for another
purpose.
If a function procedure is to return a value then put the return type (eg int, double, string
etc.) in place of ‘void’ in the function declaration. The main program must be written to
handle the value type returned (eg. The code below stores the returned result into a
double variable and not an int).